+++ auth-compat/package.json
@@ -11,8 +11,7 @@
     "build": "tsc",
     "dev": "tsx src/cli/index.ts",
     "test-validation-server": "tsx src/test-validation-server.ts",
-    "cli": "tsx src/cli/index.ts",
-    "test": "tsx src/cli/test-runner.ts"
+    "cli": "tsx src/cli/index.ts"
   },
   "dependencies": {
     "@modelcontextprotocol/sdk": "^1.0.0",
